requirements.txt
openai
langchain
streamlit
python-dotenv
PyPDF2
langchain-community
langchain-google-genai==2.0.5
google-generativeai

-e .